requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
numpy==1.26.4
Pillow==10.0.0
mtcnn==0.1.1
tensorflow==2.15.0
keras-facenet==0.3.2
scikit-learn==1.3.2
